+++ packages/adapter-oas/src/parser.ts
@@ -1,4 +1,5 @@
 import { pascalCase } from '@internals/utils'
+import { applyMacros, macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from '@kubb/ast/macros'
 import { childName, enumPropName, extractRefName, mergeAdjacentObjectsLazy } from '@kubb/ast/utils'
 import { ast } from '@kubb/core'
 import { DEFAULT_PARSER_OPTIONS, enumExtensionKeys, SCHEMA_REF_PREFIX, typeOptionMap } from './constants.ts'
@@ -330,7 +331,7 @@ export function createSchemaParser(ctx: OasParserContext, dialect: OasDialect = 
 
         const narrowedDiscriminatorNode = sharedPropertiesNode
           ? pickDiscriminatorPropertyNode(
-              ast.applyMacros(sharedPropertiesNode, [ast.macroDiscriminatorEnum({ propertyName: discriminator.propertyName, values: [discriminatorValue] })], {
+              applyMacros(sharedPropertiesNode, [macroDiscriminatorEnum({ propertyName: discriminator.propertyName, values: [discriminatorValue] })], {
                 depth: 'shallow',
               }),
               discriminator.propertyName,
@@ -373,7 +374,7 @@ export function createSchemaParser(ctx: OasParserContext, dialect: OasDialect = 
       members: unionMembers.map((s) => parseSchema({ schema: s as SchemaObject, name }, rawOptions)),
     })
 
-    return ast.applyMacros(unionNode, [ast.macroSimplifyUnion], { depth: 'shallow' })
+    return applyMacros(unionNode, [macroSimplifyUnion], { depth: 'shallow' })
   }
 
   /**
@@ -586,11 +587,11 @@ export function createSchemaParser(ctx: OasParserContext, dialect: OasDialect = 
           const resolvedChildName = childName(name, propName)
           const propNode = parseSchema({ schema: resolvedPropSchema, name: resolvedChildName }, rawOptions)
           const schemaNode = (() => {
-            const enumNameMacro = ast.macroEnumName({ parentName: name, propName, enumSuffix: options.enumSuffix })
-            const node = ast.applyMacros(propNode, [enumNameMacro], { depth: 'shallow' })
+            const enumNameMacro = macroEnumName({ parentName: name, propName, enumSuffix: options.enumSuffix })
+            const node = applyMacros(propNode, [enumNameMacro], { depth: 'shallow' })
             const tupleNode = ast.narrowSchema(node, 'tuple')
             if (tupleNode?.items) {
-              const namedItems = tupleNode.items.map((item) => ast.applyMacros(item, [enumNameMacro], { depth: 'shallow' }))
+              const namedItems = tupleNode.items.map((item) => applyMacros(item, [enumNameMacro], { depth: 'shallow' }))
               if (namedItems.some((item, i) => item !== tupleNode.items![i])) {
                 return { ...tupleNode, items: namedItems }
               }
@@ -650,7 +651,7 @@ export function createSchemaParser(ctx: OasParserContext, dialect: OasDialect = 
       const discPropName = schema.discriminator.propertyName
       const values = Object.keys(schema.discriminator.mapping)
       const enumName = name ? enumPropName(name, discPropName, options.enumSuffix) : undefined
-      return ast.applyMacros(objectNode, [ast.macroDiscriminatorEnum({ propertyName: discPropName, values, enumName })], { depth: 'shallow' })
+      return applyMacros(objectNode, [macroDiscriminatorEnum({ propertyName: discPropName, values, enumName })], { depth: 'shallow' })
     }
 
     return objectNode
