+++ packages/ast/src/macros/index.ts
@@ -1,0 +1,3 @@
+export { macroDiscriminatorEnum, macroEnumName, macroSimplifyUnion } from './builtins.ts'
+export { applyMacros, composeMacros, defineMacro } from './defineMacro.ts'
+export type { Macro } from './defineMacro.ts'
